Rename module `sys-compote` -> `compote`.
