+++ src/built-ins/export/HostRouter.js
@@ -3,7 +3,7 @@
 
 import { TreePathMap } from '@this/collections';
 import { HostUtil, IntfRequestHandler } from '@this/net-util';
-import { Names } from '@this/sys-compote';
+import { Names } from '@this/compote';
 import { BaseApplication } from '@this/sys-framework';
 import { MustBe } from '@this/typey';
