+++ src/built-ins/export/PathRouter.js
@@ -3,7 +3,7 @@
 
 import { TreePathKey, TreePathMap } from '@this/collections';
 import { DispatchInfo, IntfRequestHandler, UriUtil } from '@this/net-util';
-import { Names } from '@this/sys-compote';
+import { Names } from '@this/compote';
 import { BaseApplication } from '@this/sys-framework';
 import { MustBe } from '@this/typey';
