+++ src/compote/export/BaseNamedComponent.js
