+++ src/compote/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/sys-compote",
+  "name": "@this/compote",
   "version": "0.0.1",
   "type": "module",
   "private": true,
