+++ src/compote/private/ThisModule.js
