+++ src/sys-config/export/ClassedConfig.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseNamedConfig } from '@this/sys-compote';
+import { BaseNamedConfig } from '@this/compote';
 import { MustBe } from '@this/typey';
