+++ src/sys-config/export/SaveConfig.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseConfig } from '@this/sys-compote';
+import { BaseConfig } from '@this/compote';
 import { MustBe } from '@this/typey';
 
 import { FileServiceConfig } from '#x/FileServiceConfig';
