+++ src/sys-config/export/WarehouseConfig.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseConfig } from '@this/sys-compote';
+import { BaseConfig } from '@this/compote';
 
 import { ApplicationConfig } from '#x/ApplicationConfig';
 import { EndpointConfig } from '#x/EndpointConfig';
