+++ src/sys-config/package.json
@@ -15,7 +15,7 @@
     "@this/data-values": "*",
     "@this/fs-util": "*",
     "@this/net-util": "*",
-    "@this/sys-compote": "*",
+    "@this/compote": "*",
     "@this/typey": "*"
   }
 }
