+++ src/sys-framework/export/BaseService.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseNamedComponent } from '@this/sys-compote';
+import { BaseNamedComponent } from '@this/compote';
 import { ServiceConfig } from '@this/sys-config';
