+++ src/sys-framework/package.json
@@ -19,7 +19,7 @@
     "@this/loggy-intf": "*",
     "@this/net-protocol": "*",
     "@this/net-util": "*",
-    "@this/sys-compote": "*",
+    "@this/compote": "*",
     "@this/sys-config": "*",
     "@this/typey": "*",
     "pem": "^1.14.8"
